requirements.txt
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
pillow==12.1.1
razorpay==2.0.1
